Using molecular dynamics algorithm, allowing more individuals.
The idea is to use boxes with sides equal to neighborhood_radius so that
we restrict the area we need to check for neighbors by considering only
individuals in the neighbors boxes. As a matter of fact, we still aren't
considering cross edge contribution to theta computation, as we aren't
before.

Check this article: http://arxiv.org/pdf/1511.01451v1.pdf